refactor : 카멜 케이스노 네이밍 수정 [#09]
