+++ frontend/src/features/waiting-room/hooks/useWaitingContext.tsx
